Changed pkg header

+++ include/siri/db/query.h
@@ -56,7 +56,7 @@ typedef struct siridb_query_s
     uint8_t ref;
     uint8_t flags;
     void * data;
-    uint32_t pid;
+    uint16_t pid;
     uv_stream_t * client;
     char * q;
     char err_msg[SIRIDB_MAX_SIZE_ERR_MSG];
@@ -69,7 +69,7 @@ typedef struct siridb_query_s
 } siridb_query_t;
 
 void siridb_query_run(
-        uint32_t pid,
+        uint16_t pid,
         uv_stream_t * client,
         const char * q,
         size_t q_len,
