+++ include/siri/net/promise.h
@@ -39,7 +39,7 @@ typedef void (* sirinet_promise_cb)(
 /* the callback will always be called and is responsible to free the promise */
 typedef struct sirinet_promise_s
 {
-    uint32_t pid;
+    uint16_t pid;
     uv_timer_t * timer;
     sirinet_promise_cb cb;
     siridb_server_t * server;
